Split module from binload
